cbuild: add -B switch + config.ini option build_dir
When not specified, the behavior is the same as until now, i.e.
the builddir (where tarballs are extracted and everything is built)
will be directly inside the bldroot. However, if you do specify
the option, you can move the builddir somewhere else.

This is handy for various scenarios: putting builddir in tmpfs,
a faster SSD, or even to inspect the builddir in failed builds
with the temporary root via -t switch.