main/*: rewrite service files for better deps + autorestart
The default dinit behavior is not to automatically restart. We
want at least some services to automatically restart. Also, things
depending on dbus should be hard dependencies, not milestone,
as in practice losing dbus means losing ipc, which will break
those services anyway. For the same reason dbus should not be
smooth-recovery.