main/gnome-session: fix wayland sessions with weird shells
We patch out the horrible hack that uses bash (because it has
exec -l) to get gnome wayland sessions to source /etc/profile
in favor of a custom executable that does the same thing, but
executing gnome-session directly with $SHELL is bad because that
will only work with bourne-compatible shells. That means it will
not work with fish etc. We instead gotta execute $SHELL as a login
shell and then run the script with -c, and gnome-session will then
get executed with its shebang (/bin/sh) and things will work.