main/dinit-chimera: update
This brings support for tmpfiles management, which should make
temporary/volatile file handling a lot nicer. There might be an
apk trigger later which will perform a reload as necessary when
new tmpfiles get installed, but that is for a separate commit.

Also user services will be added later for user tmpfiles too.