main/chimera-install-scripts: new package (0.1)
This consists of 3 scripts for now - chimera-bootstrap, which
performs installations either from local or network sources,
chimera-chroot, which lets you enter a shell (or run a command)
in the target system with pseudofs mounted, and genfstab, which
generates fstabs for a mounted filesystem tree.

It replaces the scripts that are currently present in mklive.