main/apk-tools: update
This means everyone should update their host apk again. Cbuild
support will follow shortly.

This bump reintroduces support for provider-priority. It replaces
the previous priority field, which was formerly used for the
replaces-priority functionality. The reason for this is that
the field is in a section that is to be stored in index, which
is needed for provider priority but not replaces priority. The
replaces-priority is now in the other section with scripts/triggers
and so on.