main/llvm: set !int for x86_64 too
As it turns out, having it enabled also breaks the aarch64 target
when cross-compiling in specific cases, and this prevents us from
bootstrapping openjdk. I will investigate this later, but perhaps
re-test once we update to 16 and see if it's still happening,
and if yes, figure out what the bug is and fix it.