main/chimerautils: update to 13.2.1
This brings only one change but it's a large one, we switch the
tools from musl str(f|p)time to the FreeBSD libc variants, which
allows for much richer date formats. For instance, %s to print
the date in epoch format now works and so on. This unbreaks
various scripts that do support BSD date but expect the features
to work, while making the tools more useful.