main/pipewire: update to 0.3.69 and add pipewire-wireplumber
This new metapackage is automatically install_if'd with pipewire,
and depends on wireplumber. Since cbuild resolves all template deps
and builds them if not available (in order to ensure all packages
are installable), mark it virtual (even though it's not) and set
the provider to pipewire itself. This bypasses the check while
still emitting the dependency. It's an install_if package so it
does not really matter if it's always installable or not.