main/openresolv: update to 3.13.0 and change default path
The previous behavior did the right thing with /etc/resolv.conf
being a symlink, which was write into its target. As upstream
altered that to mv to /etc/resolv.conf instead, this now replaces
any potential symlink. We don't like that, so change the default
path so it always operates on the destination and leaves the
symlink alone.