cbuild: add bulk-git command
This is like bulk-pkg but for git repositories, i.e. instead of
passing a list of templates, you can pass a revision pattern,
which can be a specific commit (in which case everything after
that commit is bulk'd) or a range (which lets you specify a commit
range, i.e. with just commit alone HEAD is the implicit end).

For now there is no way to restrict the categories that will be
built, but that will be added later.