main/llvm: force lld from llvm-bootstrap as linker
Since the bootstrapping build which builds the runtimes using
the just-built toolchain is incomplete and still ends up using
regular host linker, this may result in various configure checks
failing which will besides other things result in -nodefaultlibs
being used on aarch64 instead of -nostdlib++; this will not link
the builtins in libc++abi, which will result in undefined refs.

We have a special bootstrapping linker in llvm-bootstrap so use
that unconditionally.