cbuild: default to no backtrace for pkg.error
From hooks one almost never wants to see it as it's useless (if
it isn't, the error message should be better) and in specific
cases where it's good to have it (e.g. various filesystem calls
where it's good to know which exact one failed) we can set bt
to True manually.