main/dinit-chimera: update to 0.99.1
This reduces dependency on usrmerge (for other distros), fixes
the binfmt service stop bug, and adds support for consistent
systemd-style handling of /etc/machine-id, so that dbus will
no longer have to maintain its own. It also exposes a clean way
to check for container environment as well as first boot.