main/polkit: use dinit to dbus-activate polkitd
We use a dinitctl start as an Exec of the dbus service. When
something requests polkit to be activated via dbus and it is not
running yet, the dinitctl start will take at least until it
appears on the bus (because of our readiness notification support)
and then exits. Dbus accounts for this (it does not supervise the
process) to support daemonized services, so it's fine.