cbuild: rework host/make/runtime dependency checking
This now installs all dependencies with explicit versions if
possible, which prevents a case when a template is downgraded
and a lower-version is staged (without a constraint, apk will
always prefer latest version regardless of repo order) and
something else is being built depending on it.

Repo availability is now also checked more robustly (first repo
providing the package is taken as the canonical one, as cbuild
explicitly sorts them and stage comes first).

Lastly, runtime dependencies now have their constraints checked,
so if a runtime dependency is missing and its constraint violates
whatever the template provides, we will error early instead of
attempting to build a thing that is not usable regardless.