cbuild: add bulk-raw command
This is to be used with bulk-print most of the time. It allows
the user to bulk-build multiple templates, but without support
for all the features of bulk-pkg (e.g. statusfile reading, git
and so on) and without sorting (i.e. using the order given on
cmdline).

We will use this in buildbot; bulk-print will create a label for
webui information, followed by a build, so we don't sort twice
and so that we can include previous statusfile's packages in there.