cbuild: do recursive dep checks in print-unbuilt
This way we can avoid printing templates spuriously, i.e. when
a template looks buildable but has a dependency that is not. As
print-unbuilt is not supposed to error, unparseable templates are
considered leaves and implicitly considered. Also, if a template
is in repo already and out of date, its dependencies are not
checked; this both speeds it up and allows us not to forget about
it, e.g. when a dependency is marked broken= temporarily and we
cannot get an updated version built at the time.