cbuild: move a bunch of template validation into build stage
A lot of these checks were unimportant for the purpose of parsing
the template, requiring hacks like allow_broken, ignore_errors and
so on. We can always let it parse, get rid of these, and do the
checks at a later point. For broken packages, we still need to
have the field set though (bulk builds need to know, etc).