cbuild: retry in update-check if no versions were fetched
The assumption is that every template that does not have ignored
update-check will be able to fetch a list of valid versions. As
the remote server may not always be 100% reliable, give it a few
more tries if we are unable to get any versions before printing
a warning message.