cbuild: mount cache in chroot for apk update/upgrade action
The bootstrap-update is done with in-chroot tools, and not having
the cbuild cache mounted during their operation means the cache
is not being invalidated for subsequent apk runs. That means
we may run into cbuild thinking some dependencies are out of date
and trying to rebuild them. Canceling that and running the build
again would no longer have the problem because the cache was in
the meantime invalidated by installing dependencies for the thing
being rebuilt.