cbuild: fix dep availability checking across repo categories
Previously when building a package in contrib that depends on
a package in main not present in a remote repo, we'd get an empty
string from apk search when checking in the contrib repo, which
would result in a non-empty split list (because splitting an empty
string with explicit separator yields a list with empty string
inside), which would abort the availability check. We need to
treat empty output as skippable.