cbuild: only ever write a new checksum once in prepare-upgrade
In case old source checksums had the same checksum specified
multiple times, only replace the first one. Since checksums for
URLs are typically present in the template lexically in the same
order as the URLs, it should make it behave a bit better when
using repeated placeholders, though it's not possible to fix fully.