main/dbus: source global shell profile for session bus wrapper
The most notable thing is that /etc/profile.d/flatpak.sh will
set XDG_DATA_DIRS so that dbus service files for flatpak apps
can be located. If we don't source it, the session bus will not
be aware of the dirs, so flatpak apps will not be able to
dbus-activate (as it's the session bus that launches them in
those cases).