main/efibootmgr: use simpler logic to identify whole disk drive
It seems that for a variety of reasons there is no way to reliably
map partitions to whole disk; for partitions <= 15 the whole disk
has the same major and minor 0, but beyond that it's a big unknown.
Just strip the suffix from devname instead.

Also make it clearer that the hook is not really properly supported.