main/udev: fix up tmpfiles trigger to avoid touching resolvconf
Normally within chroots etc., paths in pseudo-filesystems and so
on are not touched, but resolv.conf is a bit of a special case as
it's a symlink in /etc and is linked to /run, so the tmpfiles
trigger may touch it. This manifests as chroots suddenly losing
internet access after installing updates (e.g. during new install)
because of tmpfiles triggering.

We could special-case this to chroots but it's better to always
skip the resolv.conf config as it never really makes sense to
touch during package installs/updates, only on the subsequent
boot.

While at it, address the basename musl build fallout.