main/chrony: make the 3 minute timeout "successful"
If there is no sync after 3 minutes, we should consider chrony
started (the wait only exists so that time-sync.target can be
considered somewhat reliable, but it should not wait forever).
Still have a bigger hard timeout just in case (if that hits,
something has really gone wrong and stopping chronyd is the
right thing to do).