main/base-files: update + utilize tmpfiles, sysusers
Most of passwd/group is now generated with a sysusers snippet
and the groups have been reshuffled. This should not affect most
of existing setups as apk will not touch the files (unless they
happen to be not modified, which is highly unlikely). All of the
touched users/groups also generally only affect ephemeral files
so a reboot would fix the rest.

An exception is nobody/nogroup which we *really* want to switch
over to 65534. Do that in a shadow trigger though.