main/tzutils: merge with tzdata, remove from stage0, improve
The old package was over 7MB, now it's only 3MB. We achieve that
by properly getting rid of dupes through use of relative symlinks.
As zic does not have an option to do that and uses hardlinks by
default, we just dedup it separately afterwards. It's ugly, but
it works.