main/autoconf, main/libtool: allow for early bootstrap
These runtime depends are present in the standard bldroot so we
can avoid doing depchecks (which would trigger further builds).
This untangles autotools from the rest of the early cycle, which
allows it to be built very early.

That means we can regenerate autotools files for some bootstrap
projects without requiring pregenerated ones.