main/elfutils: fix build with clang 18
Probably some change in libc++ resulted in <stack> now being
included in standard C++ headers, which also happen to be included
from src/srcfiles.cxx. Having an executable called "stack" and
the right compiler flags would result in it attempting to include
said executable instead of the actual system stack header.

Rename the executable for build to work around this.