contrib/u-boot-imx8mq_reform2,base-reform-imx8mq: remove
The u-boot was based on a (by now outdated) downstream fork and
packaged from a binary build as building it from source with a
baremetal gcc toolchain produced broken binaries. Since we have
not been able to get help with getting that fixed and replaced
with a proper source build, and the imx8mq SOM now being basically
obsolete, remove. Somebody else is free to contribute proper
support on their own.