cbuild: implement compact exception traceback format
The full python exception traceback format is obtuse, leading to
people getting needlessly confused about what actually happened.
We still want a stack trace on many of the exceptions as it makes
it easy to locate where the error happened. Implement a custom
format that is way more compact and uses our logger's coloring
to make it look more presentable.