main/mandoc: move apropos stuff to main pkg, enable int+lto+vis+cfi
The motivation for moving `mandoc-apropos`'s contents (save the trigger)
to the main package is to allow people to do `apk add !mandoc-apropos`
but still be able to run `makewhatis`, e.g. in a cron job, instead of
using the apk trigger (because the command can take quite a bit).