main/mesa: rework dependencies
The main mesa package (which only contains some config files)
is now pulled in by most of the libraries, including lib(e)gl
and the likes.

We then make mesa-dri install_if on it. That means by default you
will get mesa gl/vulkan drivers installed automatically if there
is something that pulls in lib(e)gl(es), which is what most people
want, while still being able to remove it if needed (as it's a
soft dep)