cbuild: make maintainer global to cports clone
The maintainer field in packages for now still exits and is a
mandatory/validated field  but is ignored by package generator.
All packages get a fallback name that primarily exists to identify
the origin cports instance. That is, people are free to choose
a name for their local repos.

Update check is for now retained as is, i.e. it still reads the
maintainer field. We will address that later.

Our builders will get a `Chimera Linux <cports@chimera-linux.org>`
set on each builder.