cbuild: make maintainer global to cports clone
The maintainer field in packages for now still exits but is
ignored. All packages get a fallback name that primarily exists
to identify the origin cports instance. That is, people are free
to choose a name for their local repos.

Our builders will get a `Chimera Linux <cports@chimera-linux.org>`
set on each builder.