cbuild: fix failure report stacking
When something fails as a part of a build triggered as a dep
of another build, it's not enough to just set_failed it because
as it propagates up the stack, the set_failed will then be called
again for the parent package, resulting in the parent package
being reported instead of the actual package being built.

Since we only ever retrieve the failed template handle once
during the final print, only ever set it if it's not set already,
and erase it during retrieval (to prepare for the next potential
failure, e.g. in batch builds that don't exit on first failure).