cbuild: skip collecting repos when finding installed providers
It seems when providing --cache-dir with --from installed, apk
will include some not-installed packages in the repo as long as
they are cached. This is wrong but also when finding an installed
provider we don't really care about cache or repos at all, so
do not collect them (which also speeds it up a tad).

Strictly speaking it's harmless because we never use the mapped-to
physical provider, generally it's just informational. But it makes
the output ugly so fix it anyway.

Fixes https://github.com/chimera-linux/cports/issues/2148