main/musl: fix chungus allocations aborting instead of nullptr
This is probably not the right fix for upstream which would need
a bit more work. We modify the remap function to return false
specially for when we get a null pointer, as the only place where
we use the remap is the secondary allocator and there we can handle
such condition (and it already handles it, too).

Other scudo configurations also use this in the primary allocator
elsewhere and there I am not sure if the code is correct, so it
would need further investigation. But we rely on primary32, which
never remaps, so this is okay.

TODO: investigate this for primary64 and open an upstream fix.