cbuild: stop writing per-step logfiles
These were inherited conceptually from xbps-src but were not
super useful (I doubt anybody was ever really reading them and
sometimes they actively got in the way). We still keep the redir
code in general as it allows us to make things print to their own
pty (or a pipe when colors are disabled) instead of the actual
stdout/err of the terminal.