cbuild: implement depends list sort checking and refactor
This is only checked in the primary block, so we don't have to
concern ourselves with parsing AST for all the subpackages and
conditional manipulation and whatever we don't really care about
anyway.

Original implementation/idea by alice. The code is disabled for
now as the templates have to be reformatted.