cbuild: defer path setup until build
This allows templates to be parsed without setting up all the
paths which takes up a considerable chunk of time doing stuff we
do not even need for e.g. linting.

The side effect is that the path vars are no longer available from
subpackage bodies in templates, but only one thing used that.