cbuild: force new chroot checks when source-bootstrapping
Each of the stages does a forced chroot check to see if the
bldroot for said stage is prepared yet. If it is not, it will
revert to previous stage's bldroot and bootstrap the new package
set for that stage. We need to redo the chroot check after
reverting the stage, because the bldroot path has changed in
the meantime and the previous cached value no longer applies.