cbuild: fix recompression of manpages pointing outside mandir
This hits some specific rare cases like postgresql. It is also
somewhat tricky to handle as these links in mandir may not point
to a manpage in the same package (in postgres it does not, the
alternatives package points to main package's dir and the symlink
resolve as a dep). So we need to iterate every possible subpkg,
testing the paths as we go. Fortunately, this is narrowed down
to a slow path that is invoked only when needed.

Fixes https://github.com/chimera-linux/cports/issues/2807