cbuild: improve depsorting in bulk-pkg
The algorithm now considers packages implicitly installed in the
build container and not specified in dependencies, i.e. it behaves
as if every template specified base-cbuild in its build deps,
except if the package itself is in base-cbuild's deptree. This
prevents incorrect ordering (e.g. foo before chimerautils where
foo depends on a feature in new chimerautils release).