cbuild: optimise update-check urls for gitlab/github
the main html page is huge and takes a while (400kb+ for github); the atom url
is usually less than 1/10th the size of that and responds faster, and contains
only the versions.

we have to fix manual pattern= after this, of course, in the next commits.

also permit V and pkgname- prefixes on tags universally from these repos and
codeberg; we match 'v' in some places, and these other forms are very common, so
just allow them all.

for github, i considered using api.github.com/repos/{pn}/tags, but the
sort order is wrong- for instance if you fetch the tags on /libjpeg-turbo/
libjpeg-turbo limited to 10 (like tags.atom), the first 10 tags are the jpeg*
ones, because it's sorted not by date. that makes it not useful unless you fetch
every tag, and then the transfer is bigger than just 10 recent atom tags, which
is optimal most of the time as we don't need the full history.

if we actually need every tag, the optimal endpoint is
{pn}/info/refs?service=git-upload-pack :^)