cbuild: account for DF_1_PIE in scanelf
When this flag is set it lets us differentiate shared-libs
from executables, which may not have an interpreter in static-pie
case (so checking for lack of interpreter is not reliable).

Libraries may have an interpreter and cbuild will treat them
as an executable. That's fine because it's a rare (and harmless)
case, and not all compilers may set DD_1_PIE correctly. So while
"interpreter == executable" is not a completely robust heuristic,
it's better than the alternative.

Fixes https://github.com/chimera-linux/cports/issues/2876