cbuild: add patch_style to template, allowing `git apply`
The current default is still `patch`, but you can set it to `git`
in which case it will use `git apply` instead of `patch(1)`
command.

The Git method is somewhat stricter, does not deal with non-unified
diffs, and will not by default accept fuzzy matches. It may become
the default in the future.