main/dinit: update to git master
This is a temporary measure to get our new features in. The most
visible one is services with arguments (svcname@argvalue) for
parameterized services. The other stuff is related to being able
to listen to environment changes so we can implement autotriggered
graphical user services.

It's only a few commits on top of 0.19 so let's pretend it's still
0.19.