cbuild, */*: start installing services in usr/lib/dinit.d from now
This follows the usual pattern "if you want to modify it, copy
the file to /etc location and it will take priority". It should
help achieve no /etc in packages eventually (i.e. stateless and
potentially immutable system).

Don't rebuild anything for now as a bunch of it overlaps with
libexec switch and we can just gradually change things.