main/nyagetty: autoenable serial gettys, detect params, use args
Using args means we no longer need to generate a fixed set
of services for all possible gettys.

The autoenable behavior can be disabled in the config with
GETTY_DISABLE_CONSOLE_DETECT key. The new code results in
automatic enablement of gettys for active consoles (provided
that a service exists). Additionally, instead of hardcoding
a baud rate and other params, try detecting them first.

All this means we'll be able to remove this code from the live
initramfs, and it's a bit less work to figure out for everyone.