main/base-kernel: add new kernel.d hooks for setup
For the time being, these don't actually do anything as the
kernel packages are not set up for it. Only kernels that are
packaged for it will be matched.

The idea is that none of the kernel packages will come with any
pre or post hooks. They will also not install anything outside
of /usr/lib/modules/apk-dist.

The first kernel.d hook, 00-setup-kernels, is run before anything
else, and checks the apk-dist directory for any kernels that are
not yet "applied" into the system. If any exist, for each of these,
it performs a hardlinked recursive copy to the "proper" place -
that is, it should not take much extra space as everything is
hardlinked, except directories. The result is just a copy of the
dist kernel, in a different place.

It is expected that this directory contains boot/, which is files
that should be put in, well, /boot. The hook moves them all where
they should be.

At the end, a special marker is created in the final kernel's
directory so that we can later check that this kernel is managed
with apk and not manually or something.

The second kernel.d hook, 05-prune-kernels, is expected to run
before most others. It collects all kernels that are marked (i.e.
managed by apk) into a list and sorts this list. Then it goes over
it and filters out kernels that are either the currently booted
one, that still have apk-dist files (which means they are owned
by a package) or that are the (by default) 2 newest versions of
the same packaged series (major/minor number plus the flavor string
must match). Kernels that have not been filtered out are removed
from the system.