main/ckms: remove trigger and shared helper, consolidate
Since base-kernel now triggers on src, we can move all ckms
behavior in a single place, so it's less confusing (it's still
confusing) and can be reasoned about more easily. Also no longer
need to worry about trigger races (the triggers run in the same
order always but there is no guarantee what that order is and we
cannot control it in practice).

The new ckms kernel.d hook is transitional as it just merges the
previous two places together. However, that will change with the
new kernel management system.