main/dbus: add initial prerequisites for dbus service activation
This for now adds the flag, and updating of activation environment
through dinit-dbus. The actual activation may or may not be added
later, but it's split to allow for easier debugging.

The activation environment updating is significant though because
for most people graphical.target for user services will now trigger
automatically, without needing custom mechanisms (deactivation is
another matter, that is not handled yet, but in most cases it does
not matter because deactivation tends to be coupled with the session
bus shutting down entirely).