user/*: mass migration of rust/go things from main
A lot of these are orphaned or were drive-bys, a few are
maintained by active people. Either way they are all leaves
and random so they are a better fit for user/. As a bonus, they
get rebuilt for the latest toolchain. Some were left in main/
and those were intentional.