cbuild: route CARGO_HOME to /tmp except when preparing/vendoring
This fixes not being able to properly build multiple rust things
at once due to advisory lock over cache directory.

Once sources are vendored (and every template vendors, unless the
project comes pre-vendored) cargo is not expected to touch the
cache anymore, because it only contains crates and registry index,
both of which are irrelevant for a vendored build.