main/gsettings-desktop-schemas: use gsettings override instead of patch
The override (20_...) will be prefered to other overrides, e.g. by Phosh
(00_...) and therefore take precedence regarding the accent color and the
background. It also the preferred method by upstream for distributors to
apply own values.

The Phosh issue was only visible in gdm since Phosh's overrides only
apply if XDG_CURRENT_DESKTOP was set to "Phosh", which wasn't if run
manually.