main/linux-lts,linux-stable: disable btf typeinfo on ppc64(le)
Since clang 20, this miscompiles in some way that the runtime code
fails to parse. When enabled for modules, it prevents any modules
from loading (due to CONFIG_MODULE_ALLOW_BTF_MISMATCH not being
enabled and apparent mismatch) but when only enabled for kernel
it will still raise an error in early dmesg (will still boot).

It only seems to affect this platform. On x86_64 it works fine.