main/cryptsetup-scripts: Directly depend on lvm2-dm-initramfs-tools
This commit adds a requirement on lvm2-dm-initramfs-tools for
cryptsetup-scripts.  Normally lvm2-dm-initramfs-tools gets pulled in by
an install rule when booth lvm2-dm and initramfs-tools are installed.
This causes a problem when installing a new system inside of an chroot,
because one needs to run apk upgrade after installing cryptsetup-scripts
otherwise the lvm2-dm-initramfs-tools wont get pulled in by apk and the
user has a non-booting system.  With this change lvm2-dm-initramfs-tools
pulls in lvm2-dm via dependency and the user ends up with the correct
set of packages without needing to upgrade the system