user/ocaml-findlib: new package
The tests here are not very useful, in particular when ocamlfind is
compiled as a "custom" executable (a concatenation of the ocamlrun
bytecode interpreter and the bytecode program itself), it doesn't work,
but the tests still pass.

This isn't that important here since we're compiling a native
executable, but is worth keeping in mind.