main/dinit-chimera: update to 0.99.21
This is just fix release. It makes devmon always compiled and
always run. Previously we would skip it for containers but this
is wrong because it's a mandatory dep of early-devices.target and
is a process service that signals readiness (so it needs to do
that at least once). So instead, we add a dummy mode (also used
when compiled without udev) and run it without listening for
device events where we can't.