cbuild: implement strict service deptracking
The previous system did so based on implicit providers in repo
which is bad when bootstrapping as it may result in service
dependencies on stuff that has not been built yet. Instead,
track the dependencies the same as e.g. for .pc files and make
sure to add the providers to makedepends.