main/sqlite: enable .recover command in cli
Enabling .recover requires enabling the DBPAGE virtual table which is
dangerous and has security implications so this patch compiles the cli
with DBPAGE and the library without. The upstream binary releases of the
cli are also compiled with DBPAGE enabled, and the library without.
The ".recover" command allows you to attempt to recover data from
corrupted database files in situations where other commands like ".dump"
would not help.